feat: use Microsoft.Net.Compilers to compile
Close #11

+++ Packages/com.coffee.open-sesame-compiler/Editor/Unity.PureCSharpTests/OpenSesameCompiler.cs
@@ -1,11 +1,9 @@
-﻿using System;
-using System.Diagnostics;
+﻿using System.Text;
 using System.IO;
-using System.Linq;
 using UnityEditor;
+using System.Text.RegularExpressions;
 using UnityEditor.Scripting;
 using UnityEditor.Scripting.Compilers;
-using UnityEditor.Scripting.ScriptCompilation;
 using UnityEditor.Utils;
 using UnityEngine;
 
@@ -14,80 +12,47 @@ namespace Coffee.OpenSesameCompilers
     internal class OpenSesameCompiler : MicrosoftCSharpCompiler
     {
 #if UNITY_2019_3_OR_NEWER
+        string[] references { get { return scriptAssembly.References; } }
 		public OpenSesameCompiler(ScriptAssembly scriptAssembly, EditorScriptCompilationOptions options, string tempOutputDirectory) : base(scriptAssembly, options, tempOutputDirectory)
         {
         }
 #else
-        ScriptAssembly scriptAssembly;
-
-        public OpenSesameCompiler(ScriptAssembly scriptAssembly, MonoIsland island, bool runUpdater) : base(island, runUpdater)
+        string[] references { get { return m_Island._references; } }
+        public OpenSesameCompiler(MonoIsland island, bool runUpdater) : base(island, runUpdater)
         {
-            this.scriptAssembly = scriptAssembly;
         }
 #endif
 
         protected override Program StartCompiler()
         {
-            // Kill previous process.
+            // Install modified compiler. 
+            string installedCsc = OpenSesameInstaller.Install();
+
             var p = base.StartCompiler();
+            if (string.IsNullOrEmpty(installedCsc))
+            {
+                UnityEngine.Debug.LogWarning("<b>[OpenSesame]</b><color=orange>[Compiler]</color> OpenSesameCompiler is not installed. Instead, the default compiler will be used.");
+                return p;
+            }
+
+            // Kill previous process.
             p.Kill();
 
-#if UNITY_2019_3_OR_NEWER
-            var responsefile = assembly.GeneratedResponseFile;
-            var outputPath = assembly.FullPath;
-#else
-            // Get last responsefile.
-            var outopt = string.Format("/out:\"{0}\"", m_Island._output);
-            var responsefile = Directory.GetFiles("Temp", "UnityTempFile*")
-                    .OrderByDescending(f => File.GetLastWriteTime(f))
-                    .First(path => File.ReadAllLines(path).Any(line => line.Contains(outopt)))
-                    .Replace(Environment.CurrentDirectory + Path.DirectorySeparatorChar, "");
-            var outputPath = scriptAssembly.FullPath;
-#endif
-
-            Debug.Log("<b>[OpenSesame]</b><color=orange>[Compiler]</color> outopt: " + outopt);
-            Debug.Log("<b>[OpenSesame]</b><color=orange>[Compiler]</color> responsefile: " + responsefile);
-            Debug.Log("<b>[OpenSesame]</b><color=orange>[Compiler]</color> outputPath: " + outputPath);
-
-            // Start compiling with dotnet app
-            const string compiler = "Packages/com.coffee.open-sesame-compiler/Compiler~";
-            var psi = new ProcessStartInfo()
+            var psi = p.GetProcessStartInfo();
+            if (Application.platform == RuntimePlatform.WindowsEditor)
             {
-                Arguments = string.Format("run -p {0} -- -l {1}.log {1}", compiler, responsefile),
-                FileName = "dotnet",
-                CreateNoWindow = true
-            };
+                psi.FileName = installedCsc.Replace('/', '\\');
+            }
+            else
+            {
+                psi.FileName = Paths.Combine(EditorApplication.applicationContentsPath, "MonoBleedingEdge", "bin", "mono").Replace('\\', '/');
+                psi.Arguments = installedCsc.Replace('\\', '/') + " " + psi.Arguments.Replace("/shared ", "");
+            }
 
-            // On MacOS or Linux, PATH environmant is not correct.
-            if (Application.platform != RuntimePlatform.WindowsEditor)
-                psi.FileName = "/usr/local/share/dotnet/dotnet";
-
-            Debug.LogFormat("<b>[OpenSesame]</b><color=orange>[Compiler]</color> Execute: {0} {1}", psi.FileName, psi.Arguments);
+            Debug.LogFormat("<b>[OpenSesame]</b><color=orange>[Compiler]</color> {0} {1}", psi.FileName, psi.Arguments);
 
             var program = new Program(psi);
-            program.Start((s, e) =>
-            // Exit callback
-            {
-                if (program.ExitCode == 0)
-                {
-                    Debug.LogFormat("<b>[OpenSesame]</b><color=orange>[Compiler]</color> <color=green>Success:</color> {0}", scriptAssembly.FullPath);
-
-                    // If dll published, reimport assembly.
-                    if (!Path.GetDirectoryName(outputPath).StartsWith("Library/ScriptAssemblies"))
-                    {
-                        EditorApplication.delayCall += () =>
-                            AssetDatabase.ImportAsset(outputPath);
-                    }
-                    return;
-                }
-
-                // Show error logs.
-                foreach (var l in File.ReadAllLines(responsefile + ".log"))
-                    UnityEngine.Debug.LogError(l);
-
-                Debug.LogFormat("<b>[OpenSesame]</b><color=orange>[Compiler]</color> <color=red>Failure:</color> {0}", scriptAssembly.FullPath);
-            });
-
+            program.Start();
             return program;
         }
     }
